Backend/requirements.txt
fastapi
pydantic
python-multipart
openai
PyPDF2
pandas
python-docx
uvicorn
